🚧refactor : skeleton 폴더분리
