+++ src/components/Main/Skeleton/index.tsx
@@ -1,6 +1,6 @@
 import S from '@pages/Main/style';
-import MainHeaderSkeleton from '@pages/Main/MainHeader/MainHeader.skeleton';
-import MainBodySkeleton from '@pages/Main/MainBody/MainBodySkeleton';
+import MainHeaderSkeleton from '@components/Main/MainHeader/MainHeader.skeleton';
+import MainBodySkeleton from '@components/Main/MainBody/MainBodySkeleton';
 
 const MainPageSkeleton = () => {
   return (
