+++ src/pages/Sign/style.ts
@@ -2,29 +2,21 @@ import styled from 'styled-components';
 import { flexCenter } from '@/styles/common';
 
 const Wrapper = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
   width: 100%;
   height: 100vh;
-  ${flexCenter}
-  flex-direction: column;
   background-color: ${({ theme }) => theme.COLOR.black};
-  color: white;
-  gap: 50px;
 `;
 
 const Text = styled.div`
   color: ${({ theme }) => theme.COLOR['gray-2']};
   text-align: center;
-  font-family: Pretendard;
-  font-size: 16px;
+  font-size: 14px;
   font-style: normal;
   font-weight: 400;
   line-height: 160%;
-`;
-
-const Image = styled.div`
-  width: 800px;
-  height: 200px;
-  background-color: white;
 `;
 
 const LoginContainer = styled.div`
@@ -35,26 +27,33 @@ const LoginContainer = styled.div`
 
 interface ILoginBox {
   backgroundColor: string;
-  color: 'white' | 'black';
+  color: string;
 }
 
 const LoginBox = styled.div<ILoginBox>`
-  width: 400px;
-  height: 56px;
+  width: 320px;
+  height: 48px;
+  cursor: pointer;
   border-radius: 10px;
   font-weight: 500;
   background-color: ${({ backgroundColor }) => backgroundColor};
-  color: ${({ color }) => (color === 'white' ? 'white' : 'black')};
-  ${flexCenter}
+  color: ${({ color }) => color};
+  ${flexCenter};
   position: relative;
 `;
 
+const Content = styled.div`
+  ${flexCenter};
+  flex-direction: column;
+  color: white;
+  gap: 50px;
+`;
 const S = {
   Wrapper,
   Text,
-  Image,
   LoginContainer,
   LoginBox,
+  Content,
 };
 
 export default S;
